refactor(cli): split login service helpers
